build(theme-default): use Rslib to build the package
