+++ packages/theme-default/package.json
@@ -39,8 +39,8 @@
     "src"
   ],
   "scripts": {
-    "build": "modern build",
-    "dev": "modern build -w",
+    "build": "rslib build",
+    "dev": "rslib build -w",
     "reset": "rimraf ./**/node_modules"
   },
   "dependencies": {
@@ -60,8 +60,13 @@
     "react-syntax-highlighter": "^15.6.1"
   },
   "devDependencies": {
+    "@microsoft/api-extractor": "^7.49.2",
     "@modern-js/plugin-tailwindcss": "2.64.0",
     "@modern-js/tsconfig": "2.64.0",
+    "@rsbuild/plugin-react": "~1.1.0",
+    "@rsbuild/plugin-sass": "~1.2.0",
+    "@rsbuild/plugin-svgr": "^1.0.6",
+    "@rslib/core": "0.4.0",
     "@types/body-scroll-lock": "^3.1.2",
     "@types/hast": "3.0.4",
     "@types/jest": "~29.5.14",
