+++ packages/theme-default/postcss.config.ts
@@ -1,0 +1,9 @@
+import path from 'node:path';
+
+export default {
+  plugins: {
+    tailwindcss: {
+      config: path.join(__dirname, './tailwind.config.ts'),
+    },
+  },
+};
